+++ test/integrations/testUtils.ts
@@ -243,6 +243,9 @@ export const generateSimplifiedTrackPayload: any = (parametersOverride: any) => 
     context: removeUndefinedAndNullValues({
       externalId: parametersOverride.context.externalId,
       traits: parametersOverride.context.traits,
+      device: parametersOverride.context.device,
+      os: parametersOverride.context.os,
+      app: parametersOverride.context.app,
     }),
     anonymousId: parametersOverride.anonymousId || 'default-anonymousId',
     originalTimestamp: parametersOverride.originalTimestamp || '2021-01-03T17:02:53.193Z',
@@ -286,6 +289,7 @@ export const generatePageOrScreenPayload: any = (parametersOverride: any, eventT
       screen: {
         density: 2,
       },
+      page: parametersOverride.context.page,
       traits: parametersOverride.context.traits,
       externalId: parametersOverride.externalId,
       userAgent:
@@ -351,8 +355,9 @@ export const generateGroupPayload: any = (parametersOverride: any) => {
       screen: {
         density: 2,
       },
+      device: parametersOverride.context.device,
       traits: parametersOverride.context.traits,
-      externalId: parametersOverride.externalId,
+      externalId: parametersOverride.context.externalId,
     }),
     messageId: parametersOverride.messageId || generateAlphanumericId(36),
     session_id: parametersOverride.session_id || generateAlphanumericId(36),